Last Commit Before Breaking Change
I am going to remove the general purpose vector of World Clusters that is shared by all layers  and implement a system where each layer has its own vector of World Clusters. This will make the layer system really supreme as itshould be .
And just as alive_cluster_objects is to world clusters, the grid system also needs to have something, an easy way to access the cells that actually have data without looping through all cells in the grid.