Implemented Breaking Changes
I managed to make the layering system supreme with each layer having its world cluster. The editor has been built and subjected to minimal testing to ensure that it can at least construct and work with tile-based layers. Game systems that relied on the access to the vector of world clusters which has previously been stored by the editor to be of general access to all layers were also modified